requirements.txt
python-dotenv
pyTelegramBotAPI
langchain-openai
chromadb
langchain-text-splitters 
sentence-transformers